Refactor out `lovatAPI.getRegistrationStatus()` and `lovatAPI.getCachedRegistrationStatus()`
